improving magnetic structure factor calculation

+++ mfiles/sw_colorname.m
@@ -111,6 +111,6 @@ if isempty(RGB)
     error('sw_colorname:WrongInput','The given color name (''%s'') does not exists!',cName{1})
 end
 
-RGB = RGB';
+RGB = double(RGB');
 
 end
